+++ include/ggrpc/impl/client/client_manager.h
@@ -42,6 +42,7 @@ class ClientManager {
     virtual ~Holder() {}
     virtual void Close() = 0;
     virtual bool Expired() = 0;
+    virtual std::shared_ptr<void> Lock() = 0;
 
    protected:
     template <class T>
@@ -51,6 +52,10 @@ class ClientManager {
         sp->Close();
       }
     }
+    template <class T>
+    static std::shared_ptr<void> LockWP(std::weak_ptr<T> wp) {
+      return wp.lock();
+    }
   };
   template <class W, class R>
   struct ResponseReaderHolder : Holder {
@@ -59,6 +64,7 @@ class ClientManager {
         : wp(p) {}
     void Close() override { CloseWP(wp); }
     bool Expired() override { return wp.expired(); }
+    std::shared_ptr<void> Lock() override { return LockWP(wp); }
   };
   template <class W, class R>
   struct ReaderHolder : Holder {
@@ -66,6 +72,7 @@ class ClientManager {
     ReaderHolder(std::shared_ptr<ClientReader<W, R>> p) : wp(p) {}
     void Close() override { CloseWP(wp); }
     bool Expired() override { return wp.expired(); }
+    std::shared_ptr<void> Lock() override { return LockWP(wp); }
   };
   template <class W, class R>
   struct WriterHolder : Holder {
@@ -73,6 +80,7 @@ class ClientManager {
     WriterHolder(std::shared_ptr<ClientWriter<W, R>> p) : wp(p) {}
     void Close() override { CloseWP(wp); }
     bool Expired() override { return wp.expired(); }
+    std::shared_ptr<void> Lock() override { return LockWP(wp); }
   };
   template <class W, class R>
   struct ReaderWriterHolder : Holder {
@@ -80,14 +88,17 @@ class ClientManager {
     ReaderWriterHolder(std::shared_ptr<ClientReaderWriter<W, R>> p) : wp(p) {}
     void Close() override { CloseWP(wp); }
     bool Expired() override { return wp.expired(); }
+    std::shared_ptr<void> Lock() override { return LockWP(wp); }
   };
   struct AlarmHolder : Holder {
     std::weak_ptr<Alarm> wp;
     AlarmHolder(std::shared_ptr<Alarm> p) : wp(p) {}
     void Close() override { CloseWP(wp); }
     bool Expired() override { return wp.expired(); }
+    std::shared_ptr<void> Lock() override { return LockWP(wp); }
   };
   std::vector<std::unique_ptr<Holder>> holders_;
+  std::map<void*, Holder*> holders_map_;
 
  public:
   ClientManager(int threads) : threads_(threads) {}
@@ -118,6 +129,21 @@ class ClientManager {
                                     return holder->Expired();
                                   }),
                    holders_.end());
+    for (const auto& holder : holders_) {
+      auto it = holders_map_.find(holder->Lock().get());
+      if (it == holders_map_.end()) {
+        continue;
+      }
+      it->second = nullptr;
+    }
+    auto it = holders_map_.begin();
+    while (it != holders_map_.end()) {
+      if (it->second == nullptr) {
+        it = holders_map_.erase(it);
+      } else {
+        ++it;
+      }
+    }
   }
 
  public:
@@ -135,6 +161,7 @@ class ClientManager {
       holder->Close();
     }
     holders_.clear();
+    holders_map_.clear();
 
     SPDLOG_TRACE("ClientManager::Shutdown all client closed");
 
@@ -152,6 +179,74 @@ class ClientManager {
     }
 
     SPDLOG_TRACE("ClientManager::Shutdown finished");
+  }
+
+  template <class T>
+  using OnStateChangeFunc =
+      std::function<void(grpc::Channel*, std::chrono::system_clock::time_point,
+                         std::shared_ptr<T>, bool&)>;
+
+  template <class T>
+  struct NotifyData : Handler {
+    ClientManager* cm;
+    grpc::Channel* channel;
+    std::chrono::system_clock::time_point deadline;
+    void* p;
+    OnStateChangeFunc<T> on_notify;
+    NotifyData(ClientManager* cm, grpc::Channel* channel,
+               std::chrono::system_clock::time_point deadline, void* p,
+               OnStateChangeFunc<T> on_notify)
+        : cm(cm),
+          channel(channel),
+          deadline(deadline),
+          p(p),
+          on_notify(std::move(on_notify)) {}
+    void Proceed(bool ok) override { cm->ProceedToNotify<T>(ok, this); }
+  };
+  template <class T>
+  void NotifyOnStateChange(grpc::Channel* channel,
+                           std::chrono::system_clock::time_point deadline,
+                           std::shared_ptr<T> target,
+                           OnStateChangeFunc<T> on_notify) {
+    std::lock_guard<std::mutex> guard(mutex_);
+
+    auto it = holders_map_.find(target.get());
+    if (it == holders_map_.end()) {
+      return;
+    }
+    auto client_id = next_client_id_++;
+    auto cq = &threads_[client_id % threads_.size()].cq;
+    channel->NotifyOnStateChange(
+        channel->GetState(false), deadline, cq,
+        new NotifyData<T>(this, channel, deadline, target.get(),
+                          std::move(on_notify)));
+  }
+  template <class T>
+  void ProceedToNotify(bool ok, NotifyData<T>* p) {
+    std::lock_guard<std::mutex> guard(mutex_);
+
+    struct SafeDelete {
+      NotifyData<T>* p;
+      ~SafeDelete() { delete p; }
+    } safe_delete = {p};
+
+    auto it = holders_map_.find(p->p);
+    if (it == holders_map_.end()) {
+      return;
+    }
+    auto sp = std::static_pointer_cast<T>(it->second->Lock());
+    if (sp == nullptr) {
+      return;
+    }
+    bool repeated = false;
+    p->on_notify(p->channel, p->deadline, sp, repeated);
+    if (repeated) {
+      auto client_id = next_client_id_++;
+      auto cq = &threads_[client_id % threads_.size()].cq;
+      p->channel->NotifyOnStateChange(p->channel->GetState(false), p->deadline,
+                                      cq, p);
+      safe_delete.p = nullptr;
+    }
   }
 
   template <class W, class R>
@@ -196,7 +291,9 @@ class ClientManager {
 
     std::shared_ptr<T> p(new T(cq, std::forward<Args>(args)...),
                          [](T* p) { p->Release(); });
-    holders_.push_back(std::unique_ptr<Holder>(new H(p)));
+    auto holder = std::unique_ptr<Holder>(new H(p));
+    holders_map_.insert(std::make_pair((void*)p.get(), holder.get()));
+    holders_.push_back(std::move(holder));
     return p;
   }
 };