+++ include/ggrpc/impl/client/client_response_reader.h
@@ -39,6 +39,9 @@ class ClientResponseReader {
   detail::ReaderThunk<ClientResponseReader> reader_thunk_;
   friend class detail::ReaderThunk<ClientResponseReader>;
 
+  detail::NotifierThunk<ClientResponseReader> notifier_thunk_;
+  friend class detail::NotifierThunk<ClientResponseReader>;
+
   grpc::ClientContext context_;
   std::unique_ptr<grpc::ClientAsyncResponseReader<R>> reader_;
 
@@ -57,6 +60,10 @@ class ClientResponseReader {
   OnFinishFunc on_finish_;
   OnErrorFunc on_error_;
 
+  grpc::Alarm timeout_alarm_;
+  bool timeout_alarm_set_ = false;
+  std::chrono::milliseconds timeout_ = std::chrono::milliseconds::zero();
+
   struct SafeDeleter {
     ClientResponseReader* p;
     std::unique_lock<std::mutex> lock;
@@ -64,9 +71,9 @@ class ClientResponseReader {
     ~SafeDeleter() {
       bool del = p->release_ &&
                  p->status_ == ClientResponseReader::Status::DONE &&
-                 p->nesting_ == 0;
+                 p->timeout_alarm_set_ == false && p->nesting_ == 0;
       //SPDLOG_TRACE("del={}, release={}, status={}, timeout={}, nesting={}", del,
-      //             p->release_, (int)p->status_, p->nesting_);
+      //             p->release_, (int)p->status_, p->timeout_alarm_set_, p->nesting_);
       lock.unlock();
       if (del) {
         delete p;
@@ -77,6 +84,7 @@ class ClientResponseReader {
 
   ClientResponseReader(grpc::CompletionQueue* cq, ConnectFunc connect)
       : reader_thunk_(this),
+        notifier_thunk_(this),
         cq_(cq),
         connect_(std::move(connect)) {}
   ~ClientResponseReader() {
@@ -104,6 +112,13 @@ class ClientResponseReader {
     }
     on_error_ = std::move(on_error);
   }
+  void SetTimeout(std::chrono::milliseconds timeout) {
+    std::lock_guard<std::mutex> guard(mutex_);
+    if (status_ == Status::DONE) {
+      return;
+    }
+    timeout_ = timeout;
+  }
 
   void Connect(const W& connect) {
     std::lock_guard<std::mutex> guard(mutex_);
@@ -115,6 +130,11 @@ class ClientResponseReader {
     reader_ = connect_(&context_, connect, cq_);
     reader_->Finish(&response_, &grpc_status_, &reader_thunk_);
 
+    if (timeout_ != std::chrono::milliseconds::zero()) {
+      auto timepoint = std::chrono::system_clock::now() + timeout_;
+      timeout_alarm_.Set(cq_, timepoint, &notifier_thunk_);
+      timeout_alarm_set_ = true;
+    }
   }
 
  private:
@@ -187,6 +207,8 @@ class ClientResponseReader {
     auto st = status_;
     status_ = Status::DONE;
 
+    timeout_alarm_.Cancel();
+
     if (!ok) {
       SPDLOG_ERROR("finishing error");
       if (st == Status::CONNECTING) {
@@ -201,6 +223,29 @@ class ClientResponseReader {
     if (st == Status::CONNECTING) {
       RunCallback(d.lock, "OnFinish", on_finish_, std::move(response_),
                   std::move(grpc_status_));
+    }
+    Done(d.lock);
+  }
+  void ProceedToNotify(bool ok) {
+    SafeDeleter d(this);
+    SPDLOG_TRACE("[0x{}] ProceedToNotify: ok={} status={}", (void*)this, ok,
+                 (int)status_);
+
+    assert(status_ == Status::CONNECTING || status_ == Status::CANCELING ||
+           status_ == Status::DONE);
+
+    timeout_alarm_set_ = false;
+
+    if (!ok) {
+      Done(d.lock);
+      return;
+    }
+
+    if (status_ == Status::CONNECTING) {
+      context_.TryCancel();
+      status_ = Status::CANCELING;
+      RunCallback(d.lock, "OnError", on_error_,
+                  ClientResponseReaderError::TIMEOUT);
     }
     Done(d.lock);
   }